# bytes, strings: speed up Split{,After}Seq